Create Falcon 9 RTLS
You'll need kOS and Trajectories for this...
This is actually the second day I've been working on this finished script after 11 months of trying to make it in the first place. I know the pain of not knowing how to code an atmospheric precision lander script, so I've uploaded mine for the sake of others...enjoy! Updates coming soon.